+++ src/E1_process_results.py
@@ -612,7 +612,7 @@ def convert_components_to_dataframe(dict_values):
     df_comp = df_comp.reset_index()
 
     for i in range(len(df_comp)):
-        if df_comp.at[i, "Capacity optimization"]:
+        if df_comp.at[i, "Capacity optimization"] is True:
             df_comp.iloc[i, df_comp.columns.get_loc("Capacity optimization")] = "Yes"
         else:
             df_comp.iloc[i, df_comp.columns.get_loc("Capacity optimization")] = "No"